Rename test classes to match its subjects class' names. (#2891)
* Rename and move tests classes to match class and package names
   of its subjects under test.
* Add keys to the arrays in the data providers methods.
